build(types)!: give `MermaidConfig` required vals
Modify the `MermaidConfig` TypeScript type so that any values that are
marked as required in the JSON Schema/documentation are now also marked
as required in TypeScript.

For places where these values are truely optional
(aka in the `mermaid.initialize()` function),
the new `MermaidConfigOptions` type is used instead.

BREAKING CHANGE: The `MermaidConfig` type now has required values.
                 If you're using this type as the input to
                 `mermaid.initialize()`
                 (or other public Mermaid API functions), please
                 see the new `MermaidConfigOptions` type.