fix: move from objects to Maps
The main reason for this change was so that we wouldn't have to
validate for `constructor` and `__proto__` everytime we have to
modify or get a property on the database objects.

Another good reason would be that objects in JavaScript are generally
designed to work like dynamic structs,
where the JIT compiler would store the shape of the object and
rereference that shape everytime that object is passed around.
However, when we have a "map" with user-generated keys,
`Map`s work better because they are designed for that purpose.

Co-authored-by: Yash Singh <saiansh2525@gmail.com>
Co-authored-by: Alois Klink <alois@aloisklink.com>
Cherry-picked-from: 8a5fe53706f4ed499dfd55e10639ca216f594519